Fix access violation in ContextService.RunModule
The local variable 'item' is eligible for garbage collection after the Action in field 'item.parse' has been accessed and put on the stack, but before the call to contextSwitch.With is actually executed. After this point 'item' is not referenced in the method anymore. That means that the delegates referenced in 'item.items' can be collected before the call to 'item.parse' returns, which can result in an access violation.

This adds a call to GC.KeepAlive to make sure 'item' is not prematurely collected.